Separate out Layer into ABC and LmdbLayer.
test now pass.
